Add SendSecondFactorRegistrationEmailCommand
Used to send a SS on premise registration mail message when the identity
chose to use on-premise RA vetting.

This message was previously sent after token registration was finished
in SelfService. As by then, the only vetting option was RA vetting. Now
that we have a host of different other vetting options, it makes sense
to only send the RA mail message when appropriate.