Remove the SF id from the self-vet command
The second factor id is not available in self-service after performing
the gateway authentication to get the SF token used for self-vetting
(the authoring second factor). Only the LoA is communicated in the
SamlResponse. So removing it from this command makes sense as it can
never be filled.

See: https://www.pivotaltracker.com/story/show/184292087